NMRChemBERTa/requirements.txt
# Core ML/DL frameworks
torch
torchvision>=0.15.0
torchaudio>=2.0.0

# Transformers and tokenizers
transformers>=4.30.0
tokenizers>=0.13.0

# Chemistry libraries
rdkit>=2023.3.1
rdkit-pypi>=2022.9.5  # Alternative if conda not available

# Data processing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Utilities
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Logging and configuration
pyyaml>=6.0
wandb>=0.15.0  # For experiment tracking (optional)
tensorboard>=2.13.0  # For logging (optional)

# Development and testing
pytest>=7.3.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.3.0

# Optional: For faster data loading
pyarrow>=12.0.0

# Optional: For distributed training
accelerate>=0.20.0

# Optional: For mixed precision training
apex  # Install from source if needed